+++ ExpediaCapstone/Utilities/Helper.swift
@@ -1,0 +1,36 @@
+//
+//  Helper.swift
+//  ExpediaCapstone
+//
+//  Created by Bhargav Hegde on 23/06/23.
+//
+
+import Foundation
+
+final class Helper
+{
+    static func load<T: Decodable>(_ filename: String) -> T {
+        let data: Data
+        
+        
+        guard let file = Bundle.main.url(forResource: filename, withExtension: nil)
+        else {
+            fatalError("Couldn't find \(filename) in main bundle.")
+        }
+        
+        
+        do {
+            data = try Data(contentsOf: file)
+        } catch {
+            fatalError("Couldn't load \(filename) from main bundle:\n\(error)")
+        }
+        
+        
+        do {
+            let decoder = JSONDecoder()
+            return try decoder.decode(T.self, from: data)
+        } catch {
+            fatalError("Couldn't parse \(filename) as \(T.self):\n\(error)")
+        }
+    }
+}
