+++ ExpediaCapstone/ViewModel/ForecastsViewModel.swift
@@ -13,7 +13,7 @@ final class ForecastsViewModel: ObservableObject {
     let title = "Weather"
     let searchPrompt = "search for a city or country"
     
-    var forecasts: [Forecast] = load("forecastData.json")
+    var forecasts: [Forecast] = Helper.load("forecastData.json")
     @Published var selection = Set<Int>()
     @Published var searchText = ""
     
@@ -46,32 +46,9 @@ final class ForecastsViewModel: ObservableObject {
     
     func refresh() {
         selection = Set<Int>()
-        forecasts = load("forecastData.json")
+        forecasts = Helper.load("forecastData.json")
     }
 }
 
 
-func load<T: Decodable>(_ filename: String) -> T {
-    let data: Data
-    
-    
-    guard let file = Bundle.main.url(forResource: filename, withExtension: nil)
-    else {
-        fatalError("Couldn't find \(filename) in main bundle.")
-    }
-    
-    
-    do {
-        data = try Data(contentsOf: file)
-    } catch {
-        fatalError("Couldn't load \(filename) from main bundle:\n\(error)")
-    }
-    
-    
-    do {
-        let decoder = JSONDecoder()
-        return try decoder.decode(T.self, from: data)
-    } catch {
-        fatalError("Couldn't parse \(filename) as \(T.self):\n\(error)")
-    }
-}
+
